fix: extract `getCustomDateOptions` to utils
- change translation strings for aria-labels to be feature-independent
- adjust logic to satisfy TS
